input: touchboost: leave the governor taking care of the cpu frequency changes.
I've tried a gazillion different ways to deal with input cpu-boost events,
but in the end what really works for me and
flawlessly is leaving the governor taking care of the cpu frequency changes.
I'd rather have a global input register handler
that feeds the governor with the touch input times in usecs
and leave the boosting logic to the governor.
CAF's cpu-boost code is half broken too so
I'm not bothering with trying it again.

Change-Id: I2296bab19eb243c8550aa986d7cde13de0f8447a
Signed-off-by: franciscofranco <franciscofranco.1990@gmail.com>